major rewrite - removing the whole "nested" concept because it is unwieldly and overly general.
the new design is just as powerful but much simpler.

persistent roots are as before, but now they can only exist at the top level of a store.
instead of trying to use the persistent root mechanism itself to capture the history
of changes to a persistent root (commit, revert to version, create branch, switch branch, etc),
there will now be a command log.

unsure at the moment if this command log will be kept directly in the persistent roots
or in a separate library/location on disk.